Break apart router into multiple modules
As the project grows, it seems to make sense to break the router up from one giant module handling all of the API requests to a small base router which passes requests off to modules for each area of the Kitto API. I broke up the router based on the current endpoints defined: dashboards, events and widgets

The dashboard endpoint handles all requests to open dashboards as well as sending reload events to them. The event endpoint handles opening a new SSE connection to stream events to a client. The widget endpoint handles getting data for a specific data source or posting into a data source.
